+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Validate Middlewares
 - Logout
 - Display Post Manager for post author
+- Validate Author id before Post CRUD
 
 ## 01/06/2023
 - Implement Middleware